+++ lib/libutil/pidfile.c
@@ -1,11 +1,11 @@
-/*	$NetBSD: pidfile.c,v 1.2 1999/06/06 17:31:09 thorpej Exp $	*/
+/*	$NetBSD: pidfile.c,v 1.8 2008/04/28 20:23:03 martin Exp $	*/
 
 /*-
  * Copyright (c) 1999 The NetBSD Foundation, Inc.
  * All rights reserved.
  *
  * This code is derived from software contributed to The NetBSD Foundation
- * by Jason R. Thorpe.
+ * by Jason R. Thorpe and Matthias Scheler.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -15,13 +15,6 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *        This product includes software developed by the NetBSD
- *        Foundation, Inc. and its contributors.
- * 4. Neither the name of The NetBSD Foundation nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
  * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -38,52 +31,90 @@
 
 #include <sys/cdefs.h>
 #if defined(LIBC_SCCS) && !defined(lint)
-__RCSID("$NetBSD: pidfile.c,v 1.2 1999/06/06 17:31:09 thorpej Exp $");
+__RCSID("$NetBSD: pidfile.c,v 1.8 2008/04/28 20:23:03 martin Exp $");
 #endif
 
 #include <sys/param.h>
 #include <paths.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <string.h>
 #include <unistd.h>
 #include <util.h>
 
+static int   pidfile_atexit_done;
+static pid_t pidfile_pid;
+static char *pidfile_basename;
 static char *pidfile_path;
 
-static void pidfile_cleanup __P((void));
+static void pidfile_cleanup(void);
 
-extern const char *__progname;		/* from crt0.o */
-
-void
-pidfile(basename)
-	const char *basename;
+int
+pidfile(const char *basename)
 {
 	FILE *f;
 
-	if (pidfile_path != NULL)
-		return;
+	/*
+	 * Register handler which will remove the pidfile later.
+	 */
+	if (!pidfile_atexit_done) {
+		if (atexit(pidfile_cleanup) < 0)
+			return -1;
+		pidfile_atexit_done = 1;
+	}
 
 	if (basename == NULL)
-		basename = __progname;
+		basename = getprogname();
+
+	/*
+	 * If pidfile has already been created for the supplied basename
+	 * we don't need to create a pidfile again.
+	 */
+	if (pidfile_path != NULL) {
+		if (strcmp(pidfile_basename, basename) == 0)
+			return 0;
+		/*
+		 * Remove existing pidfile if it was created by this process.
+		 */
+		pidfile_cleanup();
+
+		free(pidfile_path);
+		pidfile_path = NULL;
+		free(pidfile_basename);
+		pidfile_basename = NULL;
+	}
+
+	pidfile_pid = getpid();
+
+	pidfile_basename = strdup(basename);
+	if (pidfile_basename == NULL)
+		return -1;
 
 	/* _PATH_VARRUN includes trailing / */
 	(void) asprintf(&pidfile_path, "%s%s.pid", _PATH_VARRUN, basename);
-	if (pidfile_path == NULL)
-		return;
+	if (pidfile_path == NULL) {
+		free(pidfile_basename);
+		pidfile_basename = NULL;
+		return -1;
+	}
 
-	if ((f = fopen(pidfile_path, "w")) == NULL)
-		return;
+	if ((f = fopen(pidfile_path, "w")) == NULL) {
+		free(pidfile_path);
+		pidfile_path = NULL;
+		free(pidfile_basename);
+		pidfile_basename = NULL;
+		return -1;
+	}
 
-	(void) fprintf(f, "%d\n", getpid());
+	(void) fprintf(f, "%d\n", pidfile_pid);
 	(void) fclose(f);
-
-	(void) atexit(pidfile_cleanup);
+	return 0;
 }
 
 static void
-pidfile_cleanup()
+pidfile_cleanup(void)
 {
-
-	if (pidfile_path != NULL)
+	/* Only remove the pidfile if it was created by this process. */
+	if ((pidfile_path != NULL) && (pidfile_pid == getpid()))
 		(void) unlink(pidfile_path);
 }
